code refector and fix issues
