+++ resources/views/admin/role/partials/style.blade.php
@@ -1,0 +1,8 @@
+<link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
+<style>
+    .select2-results__option,
+    .select2-selection__choice,
+    .select2-selection__rendered {
+        text-transform: capitalize;
+    }
+</style>
