+++ src/Requests/Permission/StorePermissionRequest.php
@@ -17,16 +17,34 @@ class StorePermissionRequest extends FormRequest
             'name' => [
                 'required',
                 'string',
+                'min:3',
                 'max:50',
                 'unique:permissions,name',
-                'regex:/^(?:[A-Za-z]+(?: [A-Za-z]+)*){3,}$/'
+                'regex:/^[A-Za-z]+(?: [A-Za-z]+)*$/'
             ]
         ];
     }
     public function messages(): array
     {
         return [
-            'name.regex' => 'The name must contain at least 3 alphabetic characters and may only contain letters and single spaces.',
+            'name.required' => 'The permission name is required.',
+            'name.string'   => 'The permission name must be a valid string.',
+            'name.min'      => 'The permission name must be at least 3 characters.',
+            'name.max'      => 'The permission name may not be greater than 50 characters.',
+            'name.unique'   => 'This permission name already exists.',
+            'name.regex'    => 'The permission name may only contain letters and spaces.',
         ];
+    }
+
+    /**
+     * Auto-format the name before validation
+     */
+    protected function prepareForValidation(): void
+    {
+        if ($this->has('name')) {
+            $this->merge([
+                'name' => ucwords(strtolower(trim($this->name)))
+            ]);
+        }
     }
 }
