+++ src/Requests/Role/UpdateRoleRequest.php
@@ -3,12 +3,30 @@
 namespace admin\admin_role_permissions\Requests\Role;
 
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
 
 class UpdateRoleRequest extends FormRequest
 {
     public function authorize(): bool
     {
         return true;
+    }
+
+    /**
+     * Trim inputs before validation.
+     */
+    protected function prepareForValidation(): void
+    {
+        $name = trim($this->input('name'));
+
+        if ($name) {
+            // Convert to lowercase first, then capitalize each word
+            $name = ucwords(strtolower($name));
+        }
+
+        $this->merge([
+            'name' => $name,
+        ]);
     }
 
     public function rules(): array
@@ -19,15 +37,21 @@ class UpdateRoleRequest extends FormRequest
                 'string',
                 'min:3',
                 'max:50',
-                'unique:roles,name,' . $this->route('role')->id,
-                'regex:/^(?:[A-Za-z]+(?: [A-Za-z]+)*){3,}$/',
+                Rule::unique('roles', 'name')->ignore($this->route('role')), // allow current role name
+                'regex:/^[A-Za-z]+(?: [A-Za-z]+)*$/', // only letters + single spaces
             ],
         ];
     }
+
     public function messages(): array
     {
         return [
-            'name.regex' => 'The name must contain at least 3 alphabetic characters and may only contain letters and single spaces.',
+            'name.required' => 'The role name is required.',
+            'name.string'   => 'The role name must be a valid string.',
+            'name.min'      => 'The role name must be at least 3 characters.',
+            'name.max'      => 'The role name may not be greater than 50 characters.',
+            'name.unique'   => 'This role name already exists.',
+            'name.regex'    => 'The role name may only contain letters and single spaces.',
         ];
     }
 }
